+++ packages/polythene-utilities/src/webfont-loader.js
@@ -1,25 +1,32 @@
-import { isServer } from "polythene-core";
+import { isServer, emit } from "polythene-core";
 
-export const addWebFont = (vendor, family, key) => {
+export const addWebFont = (vendor, config) => {
   if (isServer) return;
   if (!window.WebFontConfig) {
-    window.WebFontConfig = {};
+    const emitEvent = ({ name, familyName, fvd }) => 
+      emit("webfontloader", { name, familyName, fvd, vendor, config });
+    window.WebFontConfig = {
+      loading:      ()                => emitEvent({ name: "loading" }),
+      active:       ()                => emitEvent({ name: "active" }),
+      inactive:     ()                => emitEvent({ name: "inactive" }),
+      fontloading:  (familyName, fvd) => emitEvent({ name: "fontloading", familyName, fvd }),
+      fontactive:   (familyName, fvd) => emitEvent({ name: "fontactive", familyName, fvd }),
+      fontinactive: (familyName, fvd) => emitEvent({ name: "fontinactive", familyName, fvd }),
+    };
     (function() {
-      let wf = document.createElement("script");
+      const wf = document.createElement("script");
       wf.src = (document.location.protocol === "https:" ? "https" : "http") + "://ajax.googleapis.com/ajax/libs/webfont/1/webfont.js";
       wf.type = "text/javascript";
       wf.async = "true";
-      let s = document.getElementsByTagName("script")[0];
+      const s = document.getElementsByTagName("script")[0];
       if (s) {
         s.parentNode.insertBefore(wf, s);
       }
     })();
   }
   const vendorCfg = window.WebFontConfig[vendor] || {};
-  vendorCfg.families = vendorCfg.families || [];
-  vendorCfg.families.push(family);
-  if (key) {
-    vendorCfg.key = key;
+  if (config) {
+    Object.assign(vendorCfg, config);
   }
   window.WebFontConfig[vendor] = vendorCfg;
 };